requirements.txt
amqp==5.0.6
ansible==2.8.18
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.0.3
asgiref==3.3.1
astroid==2.5.3
asttokens==2.0.5
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.9.3
billiard==3.6.4.0
binaryornot==0.4.4
black==22.1.0
bleach==4.1.0
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
captcha==0.3
celery==5.1.0
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
click==8.0.4
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.22.5
colorama==0.4.4
command-not-found==0.3
conda==4.3.16
configparser==5.2.0
cookiecutter==1.7.2
cryptography==2.8
cssselect==1.1.0
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.5.1
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.7
django-admin-rangefilter==0.8.1
django-appconf==1.0.4
django-axes==5.14.0
django-imagekit==4.0.2
django-ipware==3.0.2
django-jazzmin==2.5.0
django-rest-framework==0.1.0
django-rq==2.4.1
django-storages==1.11.1
djangorestframework==3.13.1
dlib==19.24.0
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.0.1
executing==0.8.2
face-recognition==1.3.0
face-recognition-models==0.3.0
fake-useragent==0.1.11
fasteners==0.14.1
feedparser==6.0.2
filelock==3.7.1
frappe-bench==5.3.0
future==0.18.2
geographiclib==1.52
geopy==2.2.0
gitdb2==2.0.6
GitPython==2.1.15
gnureadline==8.0.0
gpg===1.13.1-unknown
honcho==1.0.1
html5lib==1.1
httplib2==0.14.0
huepy==1.2.1
idna==2.8
imageio==2.9.0
imageio-ffmpeg==0.4.4
importlib-metadata==1.5.0
importlib-resources==5.4.0
iniconfig==1.1.1
instabot==0.117.0
instagram-private-api==1.6.0.0
ipykernel==6.9.1
ipython==8.0.1
ipython-genutils==0.2.0
isort==5.8.0
jdcal==1.4.1
jedi==0.18.1
Jinja2==2.10.3
jinja2-time==0.2.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
keyring==18.0.1
kombu==5.1.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
linecache2==1.0.0
lockfile==0.12.2
louis==3.12.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mock==4.0.3
monotonic==1.5
more-itertools==4.2.0
MouseInfo==0.1.3
moviepy==1.0.3
mypy-extensions==0.4.3
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.10.4
notebook==6.4.8
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.4.0.46
opencv-python==4.5.5.64
openpyxl==3.0.5
packaging==20.9
pandas==1.2.4
pandocfilters==1.5.0
paramiko==2.6.0
parse==1.19.0
parso==0.8.3
pathlib==1.0.1
pathspec==0.9.0
pbr==5.4.5
pexpect==4.6.0
phonenumbers==8.12.23
pickleshare==0.7.5
pilkit==2.0
Pillow==8.2.0
pipenv==2020.11.15
platformdirs==2.5.0
pluggy==0.13.1
plyer==2.0.0
poyo==0.5.0
prettytable==0.7.2
proglog==0.1.9
prometheus-client==0.13.1
prompt-toolkit==3.0.18
protobuf==3.6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
py-notifier==0.3.0
pyaes==1.6.1
pyasn1==0.4.8
PyAutoGUI==0.9.52
pycairo==1.16.2
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pyee==8.1.0
pygame==2.0.1
PyGetWindow==0.0.9
Pygments==2.11.2
PyGObject==3.36.0
PyJWT==2.3.0
pylint==2.7.4
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.7
pyperclip==1.8.1
pyppeteer==0.2.5
pyquery==1.4.3
PyRect==0.1.4
pyRFC3339==1.1
pyrsistent==0.18.1
PyScreeze==0.1.26
PySocks==1.7.1
pytest==6.2.4
python-apt==2.0.0+ubuntu0.20.4.7
python-barcode==0.13.1
python-crontab==2.4.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-slugify==4.0.1
python3-xlib==0.15
PyTweening==1.0.3
pytz==2021.1
pywhatkit==3.9
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
qrcode==7.3.1
raven==6.10.0
redis==3.5.3
requests==2.24.0
requests-html==0.10.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
responses==0.13.3
rq==1.8.0
rsa==4.8
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
schedule==1.1.0
SecretStorage==2.3.1
semantic-version==2.8.2
Send2Trash==1.8.0
sentry-sdk==1.0.0
setuptools-rust==0.11.6
sgmllib3k==1.0.0
simplejson==3.16.0
six==1.15.0
smmap==4.0.0
smmap2==3.0.1
soupsieve==2.1
sqlparse==0.4.1
ssh-import-id==5.10
stack-data==0.2.0
systemd-python==234
Telethon==1.24.0
terminado==0.13.1
testpath==0.5.0
testresources==2.0.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.60.0
traceback2==1.4.0
traitlets==5.1.1
twilio==7.7.1
typing_extensions==4.1.1
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.25.11
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.16.2
virtualenv-clone==0.5.4
w3lib==1.22.0
wadllib==1.3.3
waitress==2.1.2
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
whitenoise==6.2.0
wikipedia==1.4.0
wordcount==1.0
wrapt==1.12.1
xkit==0.0.0
XlsxWriter==1.3.7
yahoo-fin==0.8.8
zipp==3.7.0
